Add documentation on agent sliders and bidding
This is a modest addition to the DSO+T documentation, looking
specifically at how the slider setting impacts agent bidding while
also providing an overview of the four-point bidding used in DSO+T.

Minor changes to the file structure of the documentation were made.

References for the fifth volume of DSO+T reports was also added.