Update to demo model with thermostat correction
This commit includes what appears to be a bug fix for the DSO+T HVAC
agent. The agent was not correctly translating the bid_quantity (which
I'm 95% sure is the cleared quantity) into a thermostat setting. The
correction is just a few lines of code that are clearly marked in the
file.